# expression: implement vectorized evaluation for builtinAddDateStringStringSig