PCP #12101

<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

This PR implements vectorized `builtinAddDateStringStringSig`

### What is changed and how it works?

```sh
➜ make vectorized-bench VB_FILE=Time VB_FUNC=builtinAddDateStringStringSig                    
cd ./expression && \
	go test -v -benchmem \
		-bench=BenchmarkVectorizedBuiltinTimeFunc \
		-run=BenchmarkVectorizedBuiltinTimeFunc \
		-args "builtinAddDateStringStringSig"
goos: linux
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinTimeFuncGenerated-12    	1000000000	         0.0192 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinAddDateStringStringSig-VecBuiltinFunc-12         	     378	   2892838 ns/op	  713578 B/op	   17877 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinAddDateStringStringSig-NonVecBuiltinFunc-12      	     294	   4297390 ns/op	  712456 B/op	   17867 allocs/op
PASS
ok  	github.com/pingcap/tidb/expression	3.274s
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test
 - Integration test