@SunRunAway 2019-11-15
Hi, @ekalinin 
We have a plan to use template package to generate some built-in functions, and `builtinAddDateStringStringSig` is one of them, would you like to implement them with template?

Take a look at https://docs.google.com/document/d/17XdYo_EwYaWANqbjS3i3OVSz3oLlovU2fIo1bRx4YqA/edit?usp=sharing