@qw4990 2019-11-20
> Hi, @ekalinin
> We have a plan to use template package to generate some built-in functions, and `builtinAddDateStringStringSig` is one of them, would you like to have a try and implement them with template?
> 
> Take a look at https://docs.google.com/document/d/17XdYo_EwYaWANqbjS3i3OVSz3oLlovU2fIo1bRx4YqA/edit?usp=sharing

Yes, these functions below have the same pattern, so we recommend to vectorize them by `Golang template`, and here is [an example](https://github.com/pingcap/tidb/pull/12224). 
```
builtinAddDateStringStringSig
builtinAddDateStringIntSig
builtinAddDateStringRealSig
builtinAddDateStringDecimalSig
builtinAddDateIntStringSig
builtinAddDateIntIntSig
builtinAddDateIntRealSig
builtinAddDateIntDecimalSig
```